Renamed the package from restaurant to user
